requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autograd==1.6.2
autoray==0.6.12
backcall==0.2.0
beautifulsoup4==4.12.3
bettercam==1.0.0
bleach==6.1.0
blessed==1.20.0
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.5
configparser==7.0.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
docopt==0.6.2
docstring_parser==0.16
etils==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
fastdtw==0.3.4
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
ibm-cloud-sdk-core==3.20.4
ibm-platform-services==0.55.2
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.2.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kagglehub==0.2.9
keras==3.4.1
keras-core==0.1.7
keras-cv==0.9.0
keyboard==0.13.5
kiwisolver==1.4.5
lib==4.0.0
libclang==18.1.1
lxml==4.9.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.23.0
nvidia-ml-py==12.555.43
openai==1.37.1
opencv-python==4.10.0.84
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
optree==0.12.1
packages==0.1.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pbr==6.0.0
PennyLane==0.37.0
PennyLane-qiskit==0.37.0
PennyLane_Lightning==0.37.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
pykalman==0.9.7
pymongo==3.13.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pyspnego==0.11.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
qiskit==1.1.1
qiskit-aer==0.14.2
qiskit-algorithms==0.3.0
qiskit-ibm-provider==0.11.0
qiskit-ibm-runtime==0.26.0
qiskit-machine-learning==0.7.2
redis==3.5.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-cache==0.5.2
requests_ntlm==1.3.0
rich==13.7.1
rpds-py==0.20.0
rustworkx==0.15.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sspilib==0.1.0
stack-data==0.6.3
stevedore==5.2.0
style==1.1.0
symengine==0.11.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.66
ultralytics-thop==2.0.0
update==0.0.1
urllib3==2.2.2
uuid==1.30
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarg==0.1.9
zipp==3.19.2